+++ apollo-portal/src/main/java/com/ctrip/framework/apollo/portal/spi/configuration/AuthConfiguration.java
@@ -390,6 +390,7 @@ public class AuthConfiguration {
                   this.clientRegistrationRepository)));
       http.oauth2Client();
       http.logout(configure -> {
+        configure.logoutUrl("/user/logout");
         OidcClientInitiatedLogoutSuccessHandler logoutSuccessHandler = new OidcClientInitiatedLogoutSuccessHandler(
             this.clientRegistrationRepository);
         logoutSuccessHandler.setPostLogoutRedirectUri("{baseUrl}");